Prevent quarter ends being assigned to next quarter
Previously, this would assign the end of the quarter to the start of the next quarter as the "month count" starts at 1 **and not** 0.

For example, for March, month count: 3, the previous implementation would assign it to quarter 1 ("01_apr_to_jun") as `3/3 == int(1)`. In addition, this would error for the month of December, as the quarter index 4 is not handled.